requirements.txt
Django>=5.1,<6
python-dotenv>=1.0.0
